# src/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cabros Bot API",
    "version": "0.1.0",
    "description": "Protected alerting and market-analysis APIs. Machine clients use the x-api-key header; the api-key query parameter exists only for legacy compatibility and may leak through logs. The browser admin console can use Firebase ID tokens on documented admin operations when ENABLE_FIREBASE_ADMIN_AUTH=true."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current deployment"
    }
  ],
...
  "paths": {
...
    "/api/webhook/market-scanner-alert": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Run market scanners and deliver a report",
        "operationId": "postMarketScannerAlert",
        "parameters": [
          {
            "$ref": "#/components/parameters/DryRun"
          },
          {
            "$ref": "#/components/parameters/IdempotencyKeyHeader"
          },
          {
            "$ref": "#/components/parameters/XIdempotencyKeyHeader"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/MarketScanner"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/AnalysisResult"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/IdempotencyConflict"
          },
          "502": {
            "$ref": "#/components/responses/Error"
          },
          "504": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "ApiKeyHeader": []
          },
          {
            "ApiKeyQuery": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      },
      "ApiKeyQuery": {
        "type": "apiKey",
        "in": "query",
        "name": "api-key",
        "description": "Legacy compatibility only. Prefer x-api-key because query strings may be logged."
      },
      "FirebaseBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Firebase ID token for the browser admin console. Requires an admin.viewer or admin.operator custom claim; webhook operations remain API-key-only."
      }
    },
    "parameters": {
...
      "IdempotencyKeyHeader": {
        "name": "idempotency-key",
        "in": "header",
        "schema": {
          "type": "string",
          "minLength": 1
        },
        "description": "Optional replay key. Canonical header form (x-idempotency-key is also supported)."
      },
...
      "DryRun": {
        "name": "dryRun",
        "in": "query",
        "description": "When true, analyze and return alerts without notification delivery, deduplication cache claims/writes, or signal-outcome persistence.",
        "schema": {
          "type": "boolean",
          "default": false
        }
      },
...
      "XIdempotencyKeyHeader": {
        "name": "x-idempotency-key",
        "in": "header",
        "schema": {
          "type": "string",
          "minLength": 1
        },
        "description": "Optional replay key. Alternative x-prefixed header variant."
      },
...
    },
    "requestBodies": {
...
      "MarketScanner": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/MarketScannerRequest"
            },
            "examples": {
              "default": {
                "value": {
                  "exchange": "BINANCE",
                  "timeframe": "4h",
                  "scans": [
                    "top_gainers",
                    "top_losers",
                    "volume_breakout_scanner"
                  ],
                  "limit": 5,
                  "channels": [
                    "telegram"
                  ]
                }
              },
              "ranked": {
                "value": {
                  "exchange": "BINANCE",
                  "timeframe": "4h",
                  "scans": [
                    "top_gainers"
                  ],
                  "limit": 5,
                  "ranked": true,
                  "includeMultiTimeframe": true
                }
              }
            }
          }
        }
      },
...
    },
    "responses": {
...
      "Error": {
        "description": "Request or service error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Invalid request",
              "code": "INVALID_REQUEST"
            }
          }
        }
      },
      "IdempotencyConflict": {
        "description": "The idempotency key was reused with a different request fingerprint",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Idempotency key was reused with a different payload",
              "code": "IDEMPOTENCY_CONFLICT"
            }
          }
        }
      },
...
      "AnalysisResult": {
        "description": "Analysis and delivery result",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/JsonObject"
            }
          }
        }
      },
...
    },
    "schemas": {
      "JsonObject": {
        "type": "object",
        "additionalProperties": true
      },
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "jobId": {
            "type": "string",
            "description": "Durably stored job identifier returned when queue acceptance is indeterminate."
          }
        }
      },
...
      "MarketScannerRequest": {
        "type": "object",
        "properties": {
          "exchange": {
            "type": "string",
            "default": "BINANCE"
          },
          "timeframe": {
            "type": "string",
            "default": "4h"
          },
          "scans": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "top_gainers",
                "top_losers",
                "volume_breakout_scanner",
                "smart_volume_scanner",
                "bollinger_scan"
              ]
            }
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 20,
            "default": 5
          },
          "bbw_threshold": {
            "type": "number",
            "default": 0.05
          },
          "ranked": {
            "type": "boolean",
            "default": false,
            "description": "When true, sorts each scan by actionable score and includes score/reason entries in scanResults."
          },
          "includeMultiTimeframe": {
            "type": "boolean",
            "default": false,
            "description": "When true, enriches candidates with fail-open higher-timeframe alignment from TradingView MCP. Aligned candidates receive a default +10 score modifier and counter-trend candidates a default -10 modifier. The include_multi_timeframe alias is also accepted."
          },
          "channels": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "string",
              "enum": [
                "telegram",
                "whatsapp",
                "discord"
              ]
            }
          },
          "telegramChatId": {
            "type": "string",
            "minLength": 1
          },
          "whatsappChatId": {
            "type": "string",
            "minLength": 1
          }
        }
      },
...
    }
  }
}